+++ src/Git_Updater/Plugin.php
@@ -426,9 +426,6 @@ class Plugin {
 					$transient->response[ $plugin->file ] = ( new Branch() )->set_rollback_transient( 'plugin', $plugin );
 				}
 		}
-		if ( property_exists( $transient, 'response' ) ) {
-			update_site_option( 'git_updater_plugin_updates', $transient->response );
-		}
 
 		return $transient;
 	}