+++ src/Git_Updater/Theme.php
@@ -652,9 +652,6 @@ class Theme {
 				$transient->response[ $theme->slug ] = ( new Branch() )->set_rollback_transient( 'theme', $theme );
 			}
 		}
-		if ( property_exists( $transient, 'response' ) ) {
-			update_site_option( 'git_updater_theme_updates', $transient->response );
-		}
 
 		return $transient;
 	}